actualizacion con anibal

+++ moduloClientes/views.py
@@ -1,16 +1,21 @@
 # Create your views here.
-from moduloClientes.models import Cliente
+from moduloClientes.models import Cliente, Consultas
 from django.shortcuts import render_to_response, render
 from django.utils.datastructures import MultiValueDictKeyError
 from django.db.models import Q
 from django import forms
 from django.forms.extras.widgets import SelectDateWidget
+from django.http.response import Http404
+from datetime import datetime
 
 
 def clientes(request):
-    list_clientes = Cliente.objects.all()
-    
+    list_clientes = Cliente.objects.all()    
     return render_to_response('ClientesFrontEnd/clientes.html',{'l_clienetes': list_clientes})
+
+def consultas(request):
+    list_consultas = Consultas.objects.all()
+    return render_to_response('ClientesFrontEnd/consultas.html',{'l_consultas': list_consultas})
 
 def nuevo_cliente(request):
     if request.method == 'POST':
@@ -27,25 +32,51 @@ def nuevo_cliente(request):
             form = ClienteForm()
     return render(request, 'ClientesFrontEnd/nuevoCliente.html', {'form': form})
 
+def nueva_consulta(request):
+    if request.method == 'POST':
+        form = ConsultaForm(request.POST)
+        if form.is_valid():
+            try:
+                con = Consultas(esfera=request.POST['esfera'],cilindro=request.POST['cilindro'],eje=request.POST['eje'],av=request.POST['av'],add=request.POST['add'],dp=request.POST['dp'],fecha=request.POST['fecha'],Diagnostico=request.POST['Diagnostico'],Observaciones=request.POST['Observaciones'],vista=request.POST['vista'],ojo=request.POST['ojo'],estado=request.POST['estado'])
+            except MultiValueDictKeyError:
+                con = Consultas(esfera=request.POST['esfera'],cilindro=request.POST['cilindro'],eje=request.POST['eje'],av=request.POST['av'],add=request.POST['add'],dp=request.POST['dp'],Diagnostico=request.POST['Diagnostico'],Observaciones=request.POST['Observaciones'])
+            con.save()
+            return consultas(request)
+    else:
+        if request.method != 'POST':
+            form = ConsultaForm()
+    return render(request, 'ClientesFrontEnd/nuevaConsulta.html', {'form': form})
+
 def buscar_form(request):
     return render(request, 'ClientesFrontEnd/buscar.html')
 
 def buscar_cliente(request):
     if 'q' in request.GET and request.GET['q']:
-        q = request.GET['q']
+        q = request.GET['q']        
         clientes = Cliente.objects.filter(Q(nombre__icontains=q) | Q(apellido1__icontains=q) | Q(apellido2__icontains=q))
-        return render(request, 'ClientesFrontEnd/clientes.html', {'l_clienetes': clientes})
+        if clientes :
+            return render(request, 'ClientesFrontEnd/clientes.html', {'l_clienetes': clientes})
+        else:
+            return render(request, 'ClientesFrontEnd/clientes.html', {'error': True})
+        
     else:
-        return render(request, 'ClientesFrontEnd/clientes.html', {'error': True})
+        list_clientes = Cliente.objects.all()
+        return render(request, 'ClientesFrontEnd/clientes.html', {'l_clienetes': list_clientes})
     
-def editar_cliente(request):        
-    c = Cliente.objects.get(pk=int(request.GET['q']))
+def editar_cliente(request):
+    try:     
+        c = Cliente.objects.get(pk=int(request.GET['q']))
+    except Exception:
+        raise Http404
     
     if request.method != 'POST':
-        form = ClienteForm(instance=c)
-        form.id = int(request.GET['q'])
-        if form.is_valid():
-            form.save()
+        if c:
+            form = ClienteForm(instance=c)
+            form.id = int(request.GET['q'])
+            if form.is_valid():
+                form.save()
+        else:
+            raise Http404
     else:
         if request.method == 'POST':
             form = ClienteForm(request.POST)
@@ -63,27 +94,76 @@ def editar_cliente(request):
                 return clientes(request)
     return render(request, 'ClientesFrontEnd/nuevoCliente.html', {'form': form})
 
-def eliminar_cliente(request):        
-    c = Cliente.objects.get(pk=int(request.GET['q']))
-    c.delete()
+def eliminar_cliente(request):
+    try:   
+        c = Cliente.objects.get(pk=int(request.GET['q']))
+    except Exception:
+        raise Http404
+            
+    if c:
+        c.delete()
+    else:
+        raise Http404
     return clientes(request)
+
+def eliminar_consulta(request):  
+    try:       
+        c = Consultas.objects.get(pk=int(request.GET['q']))
+    except Exception:
+        raise Http404
+    if c:
+        c.delete()
+    else:
+        raise Http404
+    return consultas(request)
 
 class ClienteForm(forms.ModelForm):
     id = forms.IntegerField(required=False)
-    cedula = forms.CharField(max_length=10,required=False)
+    cedula = forms.RegexField(max_length=10,required=False, regex=r'^[0-9]+')
     nombre = forms.CharField(max_length=30)
     apellido1 = forms.CharField(max_length=30,required=False)
     apellido2 = forms.CharField(max_length=30,required=False)
-    fecha_nacimiento = forms.DateField(widget=SelectDateWidget(),required=False)
+    fecha_nacimiento = forms.DateField(widget=SelectDateWidget(years=range(datetime.today().year-99, datetime.today().year+1 )),required=False)
     #maximo cuatro telefonos separados por coma
-    telefonos = forms.CharField(max_length=43)
+    telefonos = forms.RegexField(max_length=10, regex=r'^[0-9]+')
     direccion = forms.CharField(max_length=100)
     e_mail1 = forms.EmailField(required=False)
     e_mail2 = forms.EmailField(required=False)
-    ruc = forms.CharField(max_length=15)
+    ruc = forms.RegexField(max_length=15, regex=r'^[0-9]+')
     
     class Meta:
         model = Cliente
     
+class ConsultaForm(forms.ModelForm):
+    id = forms.IntegerField(required=False)
+  
+    VISTA_CHOICES = (
+                     ('1', 'LEJOS'),
+                     ('2', 'CERCA'),
+                     )
+    OJO_CHOICES = (
+                   ('1', 'DERECHO'),
+                   ('2', 'IZQUIERDO'),
+                   )
+    ESTADO_CHOICES = (
+                   ('1', 'PENDIENTE'),
+                   ('2', 'REALIZADA'),
+                   )
+    esfera=forms.DecimalField(max_digits=5,decimal_places=3)
+    cilindro=forms.DecimalField(max_digits=5,decimal_places=3)
+    eje=forms.CharField(max_length=30)
+    av=forms.DecimalField(max_digits=5,decimal_places=3)
+    add=forms.DecimalField(max_digits=5,decimal_places=3)
+    dp=forms.DecimalField(max_digits=5,decimal_places=3)
+    fecha=forms.DateField(widget=SelectDateWidget(),required=False)
+    Diagnostico = forms.CharField(widget=forms.Textarea,required=False)
+    Observaciones = forms.CharField(widget=forms.Textarea,required=False)
+    vista = forms.ChoiceField(choices=VISTA_CHOICES)
+    ojo = forms.ChoiceField(choices=OJO_CHOICES)      
+    estado = forms.ChoiceField(choices=ESTADO_CHOICES)
+    
+    class Meta:
+        model = Consultas
+
 class BuscarForm(forms.Form):
     query = forms.CharField()
